requirements.txt
mysql-connector-python == 8.0.27
pytelegrambotapi == 4.1.1
flask == 2.1.2